Improved the character animations
Switched to the ThirdPersonAnimationController from Shared Assets
because the blend trees for walking/running/turning/jumping were
set up nicer.
Rewrote PlayerAnimatorManager to use the new animations
Also fixed a bug where it was trying to use the PhotonView's owner
which was not valid when running in the test scene.